Unbreak CategoryMessagesTest
This test previously wasn't running because the foreach() in the data
provider was totally wrong.

Also the -details variant for fostered isn't supposed to exist, so
hardcode in an exception.

Finally, add the @coversNothing annotation since this test is just
verifying the contents of en.json, not any PHP code.

Change-Id: I7ffffcc3a910aefb082f7ff59265d3be8bc46347